+++ src/components/Tag/TagList.tsx
@@ -9,7 +9,7 @@ import "./TagList.scss";
 
 export interface TagListProps {
   children: React.ReactNode;
-  dataQA: string | null;
+  dataQA?: string | null;
 }
 
 export const TagList = ({ children, dataQA }: TagListProps) => {