+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [0.13.10]
+### Fixed
+- fix the issue race condition between logic in onBlur and onChange event inside `ValidFieldText`
+- change `dataQa` to be optional prop in `Tag` and `TagList` component
+- add `defaultOpen` prop inside `Accordion` component
+
 ## [0.13.9]
 ### Fixed
 - fix the issue that `ValidFieldText` doesn't fire the onChange function when the value is empty (case: input type numeric/float)